allow the Compiler.jl stdlib to be installed on older version of Julia (#56553)
Since JuliaLang/julia#56409, Compiler.jl as a standard library has
become available. However, for Julia versions prior to this change, even
though the stdlib can be installed via Pkg.jl, the precompilation fails
due to code compatibility issues. Consequently, when an external package
that uses the Compiler stdlib adds Compiler.jl to its Project.toml, the
package would stop working on older Julia versions.

To address this, this commit adopts the same approach as JET.jl.
Specifically, on older Julia versions, a dummy `Compiler` module is
defined, allowing dependent packages to switch between using the
Compiler.jl stdlib or the previous `Core.Compiler`. While this is a
somewhat hacky solution, it should resolve the issue for now.

Also includes a change to include `ssair/show.jl` in the context of
`Compiler` to ensure that stale precompilation caches are not used.

And as a result this commit bumps the version of the Compiler.jl
standard library.