inference: complete the inference even for recursive cycles (#56551)
We care more and more now that our inference graph exactly matches the
callgraph, even in cases like this where we can easily prove that the
inference graph is simpler than the full callgraph. However, given when
the optimizer runs, it expects this information to be available and
valid as soon as the cycles are completed.